pyproject.toml
[tool.poetry]
name = "document-extraction"
version = "0.1.0"
description = ""
authors = ["Raymond <akulubala@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.12"}
llama-index = "^0.12.26"
httpx = {extras = ["socks"], version = "^0.28.1"}
django-environ = "^0.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
